Add support for dumping SoundEffect to WAVE
This adds a very hardcoded WAVE file dumper.
main() will now dump the input file to a wave file with the same name.
Note that it doesn't even sanity check that the input file was a
SoundEffect!
TODO: Write a more generic RIFF file output module
TODO: Make it dump to the right file type depending on object type